`ServiceCollectionExtensions` is not a name a library should claim
`Quartz.ServiceCollectionExtensions` held eighteen members, seven of which
extended `IServiceCollection` and eleven `IQuartzBuilder`, under the single most
common helper-class name in .NET. An application with a `ServiceCollectionExtensions`
of its own and a `using Quartz;` got CS0104 in a file with nothing to do with
Quartz — and a reader looking for `AddJob` had no reason to look in a class named
after `IServiceCollection`.

The `AddQuartz*` members join `AddQuartzHostedService` on
`QuartzServiceCollectionExtensions`, as two halves of one partial class; the
`IQuartzBuilder` members become `QuartzBuilderExtensions`. Extension-form call
sites — every call anyone writes — are untouched.
